+++ packages/components/src/code-editor/code-editor.css.ts
@@ -6,11 +6,6 @@ export const wrapCls = style({
     width: '100%',
 });
 
-globalStyle(`${wrapCls} .cm-editor`, {
-    height: '100px',
-    outline: 'none',
-});
-
 globalStyle(`${wrapCls}.is-bordered .cm-editor`, {
     border: '1px solid #cfd0d3',
     borderRadius: '4px',
@@ -19,6 +14,7 @@ globalStyle(`${wrapCls}.is-bordered .cm-editor`, {
 globalStyle(`${wrapCls}.is-bordered .cm-editor.cm-focused`, {
     border: '1px solid #5384ff',
     boxShadow: '0 0 0 2px #dde6ff',
+    outline: 'none',
 });
 
 export const fullScreenWrapperCls = style({});