+++ packages/plugin-code/src/code/edit/query-edit/rest/params.tsx
@@ -67,7 +67,7 @@ export default defineComponent({
 
                 <ContentItem label="数据转换" labelStyle="width: 72px" v-slots={{
                     content: () => {
-                        return <div style="width: 100%">
+                        return <div style="width: 0; flex: 1;">
                             <FCheckbox v-model={props.codeItem.enableTransformer}>开启数据转换</FCheckbox>
                             {!props.codeItem.enableTransformer && <p class={tipCls}>开始数据转换，将请求数据转换成不同的格式</p>}
                             {props.codeItem.enableTransformer && <CodeEditor extensions={[javascript()]} documentModel={props.documentModel} doc={props.codeItem.transformer} changeDoc={changeTransformer} /> }