+++ packages/plugin-schema/src/index.css.ts
@@ -1,4 +1,4 @@
-import { globalStyle, style } from '@vanilla-extract/css';
+import { style } from '@vanilla-extract/css';
 
 export const iconCls = style({
     cursor: 'pointer',
@@ -19,12 +19,4 @@ export const codeCls = style({
 export const actionCls = style({
     padding: '6px 0',
     textAlign: 'right',
-});
-
-globalStyle(`${codeCls} .cm-editor`, {
-    border: '1px solid #cfd0d3',
-    borderRadius: 0,
-    borderRight: 'none',
-    borderLeft: 'none',
-    height: '100%',
 });