+++ packages/renderer/src/code-impl/javascript-function.ts
@@ -1,10 +1,8 @@
-import { markComputed, markShallowReactive } from '@harrywan/letgo-common';
 import type { IJavascriptFunction } from '@harrywan/letgo-types';
-import type { IJavascriptFunctionImpl } from '@harrywan/letgo-designer';
 import { CodeType } from '@harrywan/letgo-types';
 
 // 解析执行
-export class JavascriptFunctionImpl implements IJavascriptFunctionImpl {
+export class JavascriptFunctionLive {
     id: string;
     type: CodeType.JAVASCRIPT_FUNCTION = CodeType.JAVASCRIPT_FUNCTION;
     ctx: Record<string, any>;
@@ -12,32 +10,14 @@ export class JavascriptFunctionImpl implements IJavascriptFunctionImpl {
     funcBody: string;
 
     constructor(data: IJavascriptFunction, deps: string[], ctx: Record<string, any>) {
-        markShallowReactive(this, {
-            id: data.id,
-            funcBody: data.funcBody,
-        });
-        markComputed(this, ['view']);
+        this.id = data.id;
+        this.funcBody = data.funcBody;
         this.ctx = ctx;
         this.deps = deps;
     }
 
     changeDeps(deps: string[]) {
         this.deps = deps;
-    }
-
-    get view() {
-        return {
-            id: this.id,
-            funcBody: this.funcBody,
-        };
-    }
-
-    changeId(id: string) {
-        this.id = id;
-    }
-
-    changeContent(content: Partial<IJavascriptFunction>) {
-        Object.assign(this, content);
     }
 
     async trigger() {