+++ packages/simulator-renderer/src/simulator-view/renderer-view.tsx
@@ -1,11 +1,11 @@
 import { computed, defineComponent, h, onUnmounted, provide, watch } from 'vue';
 import type { PropType } from 'vue';
-import type { JavascriptFunctionImpl } from '@harrywan/letgo-renderer';
 import LowCodeRenderer from '@harrywan/letgo-renderer';
 import { type CodeItem, CodeType, type IPublicTypeComponentInstance, type IPublicTypeNodeSchema } from '@harrywan/letgo-types';
 import type { DocumentInstance, VueSimulatorRenderer } from '../interface';
 import { BASE_COMP_CONTEXT } from '../constants';
-import { useCodesInstance } from '../context/code-impl';
+import type { JavascriptFunctionImpl } from '../code-impl/javascript-function';
+import { useCodesInstance } from '../code-impl/code-impl';
 import { useContext } from '../context/context';
 import { host } from '../host';
 import { Hoc } from './hoc';
@@ -62,7 +62,7 @@ export default defineComponent({
             }),
             code.value.onCodeIdChanged((id: string, preId: string) => {
                 changeCodeInstanceId(id, preId);
-                executeCtx[id] = codesInstance[id];
+                executeCtx[id] = executeCtx[preId];
                 delete executeCtx[preId];
             }),
         );