+++ packages/designer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @harrywan/letgo-designer
 
+## 0.0.0-beta.11
+
+### Patch Changes
+
+- update component style
+- Updated dependencies
+  - @harrywan/letgo-common@0.0.0-beta.11
+  - @harrywan/letgo-editor-core@0.0.0-beta.11
+
 ## 0.0.0-beta.10
 
 ### Patch Changes