+++ packages/engine-plugin/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @harrywan/letgo-engine-plugin
 
+## 0.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @harrywan/letgo-common@0.0.0-beta.11
+  - @harrywan/letgo-designer@0.0.0-beta.11
+  - @harrywan/letgo-editor-core@0.0.0-beta.11
+  - @harrywan/letgo-editor-skeleton@0.0.0-beta.11
+
 ## 0.0.0-beta.10
 
 ### Patch Changes