+++ packages/plugin-schema/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @harrywan/letgo-plugin-schema
 
+## 0.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @harrywan/letgo-components@0.0.0-beta.11
+  - @harrywan/letgo-designer@0.0.0-beta.11
+  - @harrywan/letgo-engine-plugin@0.0.0-beta.11
+
 ## 0.0.0-beta.10
 
 ### Patch Changes